// src/types.ts
export enum SectionId {
  Home = 'fold',
  About = 'about',
  Guide = 'guide',
  Partner = 'partner',
  GameRule = 'game-rule',
  LiveEvents = 'live-events',
  OtherEvents = 'other-events'
}

export enum WorkshopType {
  Healing = '療癒系',
  Adventure = '冒險系'
}

export enum ActiveItem {
  ExternalVillage = 'ExternalVillage',
  TailorShop = 'TailorShop',
  BookRentalStore = 'BookRentalStore',
  Cinema = 'Cinema',
  StoryTelling = 'StoryTelling',
  RebornLobby = 'RebornLobby'
}